+++ libs/i18n/src/lib/translations/translations_ms.ts
@@ -328,7 +328,8 @@ export default {
         inputGlyphAriaLabel: 'Pilih Pilihan',
         inputIconTitle: 'Pilih Pilihan',
         mobileShowAllItemsButton: 'Tunjukkan semua item',
-        mobileShowSelectedItemsButton: 'Tunjukkan item dipilih'
+        mobileShowSelectedItemsButton: 'Tunjukkan item dipilih',
+        invalidEntryError: 'Invalid Entry'
     },
     platformTextarea: {
         counterMessageCharactersOverTheLimitSingular: '1 aksara melebihi had',