+++ libs/i18n/src/lib/translations/translations_ru.ts
@@ -333,7 +333,8 @@ export default {
         inputGlyphAriaLabel: 'Выбрать опции',
         inputIconTitle: 'Выбрать опции',
         mobileShowAllItemsButton: 'Показать все элементы',
-        mobileShowSelectedItemsButton: 'Показать выбранные элементы'
+        mobileShowSelectedItemsButton: 'Показать выбранные элементы',
+        invalidEntryError: 'Invalid Entry'
     },
     platformTextarea: {
         counterMessageCharactersOverTheLimitSingular: 'Превышен лимит на 1 символ',